+++ src/main/java/eu/zunix/ore_and_alloy/OreAndAlloy.java
@@ -16,6 +16,7 @@ import eu.zunix.ore_and_alloy.item.PeriodicTooltip;
 import eu.zunix.ore_and_alloy.registry.ModCreativeTabs;
 import eu.zunix.ore_and_alloy.registry.ModFluids;
 import eu.zunix.ore_and_alloy.registry.ModOreBlocks;
+import eu.zunix.ore_and_alloy.registry.ModRawBlocks;
 import eu.zunix.ore_and_alloy.registry.ModStorageBlocks;
 
 
@@ -29,6 +30,7 @@ public class OreAndAlloy {
         ModFluids.FLUIDS.register(modEventBus);
         ModFluids.BLOCKS.register(modEventBus);
         ModOreBlocks.BLOCKS.register(modEventBus);
+        ModRawBlocks.BLOCKS.register(modEventBus);
         ModStorageBlocks.BLOCKS.register(modEventBus);
         ModItems.ITEMS.register(modEventBus);
         ModCreativeTabs.CREATIVE_TABS.register(modEventBus);